chore: split tests in two files
